backend/requirements/default.txt
aiohttp==3.9.2
alembic==1.10.4
asyncpg==0.27.0
atlassian-python-api==3.37.0
beautifulsoup4==4.12.2
celery==5.3.4
chardet==5.2.0
dask==2023.8.1
ddtrace==2.6.5
distributed==2023.8.1
fastapi==0.109.2
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==5.0.0
filelock==3.12.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
# GPT4All library has issues running on Macs and python:3.11.4-slim-bookworm
# will reintroduce this when library version catches up
# gpt4all==2.0.2
httpcore==0.16.3
httpx[http2]==0.23.3
httpx-oauth==0.11.2
huggingface-hub==0.20.1
jira==3.5.1
langchain==0.1.9
litellm==1.27.10
llama-index==0.9.45
Mako==1.2.4
msal==1.26.0
nltk==3.8.1
docx2txt==0.8
Office365-REST-Python-Client==2.5.4
oauthlib==3.2.2
openai==1.3.5
openpyxl==3.1.2
playwright==1.41.2
psutil==5.9.5
psycopg2-binary==2.9.9
pycryptodome==3.19.1
pydantic==1.10.7
PyGithub==1.58.2
python-dateutil==2.8.2
python-gitlab==3.9.0
python-pptx==0.6.23
pypdf==3.17.0
pytest-mock==3.12.0
pytest-playwright==0.3.2
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.7
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2  # This pulls in py which is in CVE-2022-42969, must remove py from image
rfc3986==1.5.0
# need to pin `safetensors` version, since the latest versions requires
# building from source using Rust
rt==3.1.2
safetensors==0.3.1
sentence-transformers==2.2.2
slack-sdk==3.20.2
SQLAlchemy[mypy]==2.0.15
starlette==0.36.3
supervisor==4.2.5
tensorflow==2.15.0
tiktoken==0.4.0
timeago==1.0.16
torch==2.0.1
torchvision==0.15.2
transformers==4.36.2
uvicorn==0.21.1
zulip==0.8.2
hubspot-api-client==8.1.0
zenpy==2.0.41
GitPython==3.1.42
